requirements.txt
streamlit==1.26.1
streamlit-option-menu==0.3.6
PyPDF2
langchain-community
sentence-transformers
faiss-cpu
groq
openai==1.3.7
python-dotenv
